updated route and db with delete
